requirements.txt
aiohttp==3.8.3
aiosignal==1.3.1
async-timeout==4.0.2
attrs==22.1.0
charset-normalizer==2.1.1
discord.py==2.1.0
frozenlist==1.3.3
idna==3.4
multidict==6.0.3
peewee==3.15.4
python-dotenv==0.21.0
yarl==1.8.2
twitchio